requirements.txt
numpy==2.2.6
opencv-python==4.12.0.88
pillow==12.0.0
requests==2.32.3
boto3==1.35.36
python-dotenv==1.0.0
